[refactor] dashboard MVC controller 분리 #98
